@@ -4,14 +4,15 @@
   "description": "A discord bot for MapleStory",
   "scripts": {
     "start": "node dist/index.js",
+    "node": "node",
     "build": "tsc",
+    "clean": "rm -rf dist",
     "formatAll": "prettier --write \"src/**/*.ts\"",
     "format": "prettier --write",
     "lint": "eslint . --ext .js,.ts",
-    "watch-nodemon": "nodemon --watch dist --inspect=0.0.0.0:9229 --nolazy",
+    "watch-nodemon": "nodemon -L --watch dist --inspect=0.0.0.0:9229 --nolazy",
     "watch-tsc": "tsc && tsc -w",
-    "test": "mocha --reporter dot --require ts-node/register test/**/*.ts",
-    "typeorm": "ts-node ./node_modules/typeorm/cli.js"
+    "test": "mocha --reporter dot --require ts-node/register test/**/*.ts"
   },
   "author": "Andy Chen <telunchen@hotmail.ca> (https://github.com/telunc)",
   "bugs": {
@@ -39,46 +40,59 @@
   "license": "MIT",
   "devDependencies": {
     "@types/async": "^3.2.3",
-    "@types/chai": "^4.2.12",
-    "@types/d3-array": "^2.0.0",
-    "@types/lodash": "^4.14.160",
+    "@types/bull": "^3.14.4",
+    "@types/chai": "^4.2.13",
+    "@types/config": "0.0.36",
+    "@types/d3-array": "^2.1.0",
+    "@types/ioredis": "^4.17.5",
+    "@types/lodash": "^4.14.162",
     "@types/mocha": "^8.0.3",
     "@types/moment-duration-format": "^2.2.2",
-    "@types/node": "^14.6.1",
+    "@types/node": "^14.11.8",
     "@types/numeral": "0.0.28",
-    "@types/socket.io-client": "^1.4.33",
+    "@types/socket.io-client": "^1.4.34",
     "@types/string-similarity": "^3.0.0",
     "@types/validator": "^13.1.0",
-    "@types/ws": "^7.2.6",
-    "@typescript-eslint/eslint-plugin": "^3.10.1",
-    "@typescript-eslint/parser": "^3.10.1",
+    "@types/ws": "^7.2.7",
+    "@typescript-eslint/eslint-plugin": "^4.4.0",
+    "@typescript-eslint/parser": "^4.4.0",
     "chai": "^4.2.0",
-    "eslint": "^7.7.0",
-    "eslint-config-prettier": "^6.11.0",
+    "eslint": "^7.11.0",
+    "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "mocha": "^8.1.2",
+    "mocha": "^8.1.3",
     "nodemon": "^2.0.4",
-    "prettier": "^2.1.1",
+    "prettier": "^2.1.2",
     "ts-node": "^9.0.0",
-    "typescript": "^4.0.2"
+    "typescript": "^4.0.3"
   },
   "dependencies": {
     "@discordjs/opus": "^0.3.2",
+    "@sentry/node": "^5.27.1",
+    "@sentry/tracing": "^5.27.1",
     "async": "^3.2.0",
-    "d3-array": "^2.6.0",
+    "bufferutil": "^4.0.1",
+    "bull": "^3.18.0",
+    "config": "^3.3.2",
+    "d3-array": "^2.8.0",
     "discord.js": "^12.3.1",
-    "dotenv": "^8.2.0",
-    "googleapis": "^59.0.0",
+    "entities": "^2.0.3",
+    "erlpack": "^0.1.3",
+    "googleapis": "^61.0.0",
+    "got": "^11.7.0",
+    "ioredis": "^4.17.3",
     "lodash": "^4.17.20",
-    "moment": "^2.27.0",
+    "moment": "^2.29.1",
     "moment-duration-format": "^2.3.2",
     "numeral": "^2.0.6",
-    "pg": "^8.3.3",
+    "pg": "^8.4.1",
     "prism-media": "^1.2.2",
     "reflect-metadata": "^0.1.13",
-    "socket.io-client": "^2.3.0",
+    "socket.io-client": "^2.3.1",
+    "sodium": "^3.0.2",
     "string-similarity": "^4.0.2",
-    "typeorm": "^0.2.25",
-    "ytdl-core": "^3.2.2"
+    "typeorm": "^0.2.28",
+    "utf-8-validate": "^5.0.2",
+    "ytdl-core": "^3.3.0"
   }
 }
