@@ -1,30 +1,29 @@
 import { Color } from "../lib";
-import { Time } from "../maplestory/Time";
 import { DiscordData, Command, CommandExample } from "./base";
 
-export class Invasion extends Command {
+export class Invite extends Command {
   constructor() {
-    super("invasion", Color.PINK);
+    super("invite", Color.BLUE);
   }
 
   public description() {
-    return "Time until the next Kritias Invasion.";
+    return "Generate invite link.";
   }
 
   public customExamples(): CommandExample[] {
     return [
       {
         cmd: this.fullName,
-        explain: "show time until the next Kritias Invasion",
+        explain: "generate invite link",
       },
     ];
   }
 
   public async continue(discord: DiscordData) {
-    const embed = this.embed().addField(
-      "Kritias Invasion in",
-      Time.formatDuration(Time.invasion)
-    );
+    const link = await discord.guild.client.generateInvite(["ADMINISTRATOR"]);
+    const embed = this.embed();
+    embed.setTitle("Invite Link");
+    embed.setDescription(link);
     discord.channel.send(embed).catch(console.error);
   }
 }
