@@ -1,9 +1,9 @@
-import { Color, Discord, Md } from "../lib";
+import { Color } from "../lib";
 import { DiscordData, Command, CommandExample } from "./base";
 
 export class Join extends Command {
   constructor() {
-    super("join", Color.MUSIC);
+    super("join", Color.PURPLE);
   }
 
   public async continue(data: DiscordData) {
@@ -11,7 +11,7 @@ export class Join extends Command {
   }
 
   public description() {
-    return "Join the member's voice channel.";
+    return "Join voice channel.";
   }
 
   public customAliases() {
@@ -22,45 +22,29 @@ export class Join extends Command {
     return [
       {
         cmd: this.fullName,
-        explain: "make the bot join into your voice channel",
+        explain: "make the bot join into the member's voice channel",
       },
     ];
   }
 
   public async join(discord: DiscordData, forceMove = false) {
     const botVoice = discord.bot.voice;
-    const memberVoice = discord.member.voice;
-    if (!memberVoice.channel) {
+    const rqtrVoice = discord.member.voice;
+    if (!rqtrVoice.channel) {
       discord.message.reply("join a voice channel first!");
       return;
     }
-    if (!botVoice.connection) {
-      return await this.attemptJoin(memberVoice.channel, discord);
-    }
-    if (botVoice.connection.channel.id === memberVoice.channel.id) {
-      return botVoice.connection;
-    }
-    if (forceMove) {
-      return await this.attemptJoin(memberVoice.channel, discord);
-    }
-  }
-
-  private async attemptJoin(
-    channel: Discord.VoiceChannel,
-    discord: DiscordData
-  ) {
-    try {
-      const connection = await channel.join();
+    if (!botVoice.connection || forceMove) {
+      const connection = await rqtrVoice.channel.join();
+      connection.on("error", console.error);
+      await connection.voice.setSelfDeaf(true);
       discord.channel.send(
-        this.embed().setDescription(`Joined ${Md.bld(connection.channel.name)}`)
+        this.embed().setDescription(`Joined ${rqtrVoice.channel}`)
       );
       return connection;
-    } catch (e) {
-      console.error(e);
-      discord.channel.send(
-        this.embed().setDescription("Unable to join " + Md.bld(channel.name))
-      );
-      return;
+    }
+    if (botVoice.connection.channel.id === rqtrVoice.channelID) {
+      return botVoice.connection;
     }
   }
 }
