@@ -1,4 +1,4 @@
-import { _, Color, Md } from "../lib";
+import { Color, Md } from "../lib";
 import { guildMaster } from "../manager";
 import { Prefix as PrefixEntity } from "../db/entity";
 import { Prefix as PrefixModel } from "../model";
@@ -9,17 +9,19 @@ export class Prefix extends Command {
   private arg: Argument;
   private space: Option;
   constructor() {
-    super("prefix", Color.GENERAL);
-    this.space = new Option("-s", "include a space after");
+    super("prefix", Color.BLUE, true);
+    this.space = new Option("-s", "include a space");
     this.arg = new Argument({ name: "CONTENT" });
     this.addCustomArg(this.arg);
     this.addCustomOptions(this.space);
   }
 
   public description() {
-    return `Change the current prefix ${Md.pre(
-      this.prefix.toString()
-    )} to a new one. If nothing is provided, the current one stays the same.`;
+    return `Change the current prefix ${Md.pre(this.prefix)} to a new one.`;
+  }
+
+  public briefDescription() {
+    return "Customize prefix";
   }
 
   public customArgDescriptions() {
@@ -41,62 +43,49 @@ export class Prefix extends Command {
         cmd: `${this.fullName} ${this.space} ${prefix}`,
         explain: `change prefix to ${Md.pre(
           prefix
-        )} with a following space. E.g., ${Md.pre(prefix + " help")}`,
+        )} with a space. E.g., ${Md.pre(prefix + " help")}`,
       },
     ];
   }
 
   public async continue(discord: DiscordData) {
     if (!this.arg.input) {
-      discord.channel.send(this.currentPrefixEmbed()).catch(console.error);
       return;
     }
 
     const includeSpace = this.space.enabled;
-    const prefix = new PrefixModel(
-      discord.guild.id,
-      this.arg.input,
-      includeSpace
-    );
+    const prefix = new PrefixModel(this.arg.input, includeSpace);
 
     if (prefix.toString() === this.prefix.toString()) {
-      discord.channel.send(this.currentPrefixEmbed()).catch(console.error);
       return;
     }
 
     const embed = this.embed();
 
     try {
-      await this.save(prefix);
-      this.prefix = prefix;
-      guildMaster.get(discord.guild.id).prefix = prefix;
-      const title = `Prefix set to ${Md.pre(prefix)}`;
-      embed.setTitle(title);
+      await this.save(discord.guild.id, prefix);
+      embed.setTitle(`Prefix is now set to ${Md.pre(prefix)}`);
       embed.setDescription(
-        `${_.capitalize(this.hasSpace())}. E.g., ${Md.pre(this.fullName)}`
+        `${this.detail()}. E.g., ${Md.pre(`${prefix}help`)}`
       );
     } catch (e) {
-      embed.setDescription("Unable to save new prefix");
+      embed.setDescription("Unable to save prefix");
       console.error(e);
     }
     discord.channel.send(embed).catch(console.error);
   }
 
-  private async save(prefix: PrefixModel) {
+  private async save(guildId: string, prefix: PrefixModel) {
     const newPrefix = new PrefixEntity();
-    newPrefix.id = prefix.guildId;
+    newPrefix.id = guildId;
     newPrefix.content = prefix.content;
     newPrefix.space = prefix.space;
     await getRepository(PrefixEntity).save(newPrefix);
+    guildMaster.get(guildId).prefix = prefix;
+    this.prefix = prefix;
   }
 
-  private currentPrefixEmbed() {
-    return this.embed().setDescription(
-      `Current prefix is ${Md.pre(this.prefix)} ${this.hasSpace()}`
-    );
-  }
-
-  private hasSpace() {
+  private detail() {
     return (this.prefix.space ? "with a " : "without any ") + "space";
   }
 }
