@@ -1,8 +1,10 @@
 import { ShardingManager } from "discord.js";
 import { env } from "./lib";
 
+process.on("unhandledRejection", console.error);
+
 const manager = new ShardingManager("./dist/bot.js", {
-  token: env.discord,
+  token: env.discord.secret,
 });
 
 manager.on("shardCreate", (shard) =>
