@@ -6,9 +6,21 @@ import * as D3Array from "d3-array";
 import moment = require("moment");
 import momentDurationFormatSetup = require("moment-duration-format");
 import * as numeral from "numeral";
+import * as entities from "entities";
+import { youtube_v3 as YouTube } from "googleapis";
 
 momentDurationFormatSetup(moment);
 
-export { _, Discord, D3Array, async, moment, numeral, stringSimilarity };
+export {
+  _,
+  Discord,
+  D3Array,
+  YouTube,
+  async,
+  entities,
+  moment,
+  numeral,
+  stringSimilarity,
+};
 export { env } from "./env";
-export { Color, DiscordEmbed, Markdown as Md } from "./util";
+export { Color, Markdown as Md } from "./util";
