@@ -1,4 +1,9 @@
-import { Prefix, Queue } from ".";
+import { Prefix, Queue, SongData } from ".";
+
+export interface GuildData {
+  id: string;
+  songs: SongData[];
+}
 
 export class Guild {
   public static readonly LIST_MAX = 200;
@@ -9,8 +14,15 @@ export class Guild {
 
   constructor(id: string) {
     this.id = id;
-    this.prefix = new Prefix(id);
+    this.prefix = new Prefix();
     this.queue = new Queue();
     this.cooldowns = new Set();
+  }
+
+  public toData(): GuildData {
+    return {
+      id: this.id,
+      songs: this.queue.songs.map((song) => song.toData()),
+    };
   }
 }
